requirements.txt
streamlit
python-dotenv
openai
requests
beautifulsoup4
langchain
langchain-community
langchain-openai
langchain-groq
sentence-transformers
faiss-cpu
PyPDF2
python-docx
pypdf
docx2txt
streamlit_webrtc
bs4
fastapi
tf-keras
langchain_huggingface
uvicorn
fastapi
pydantic
python-multipart
requests_html
lxml_html_clean